Refactor user and activity edit

+++ src/main/java/ru/forumcalendar/forumcalendar/service/base/BaseUserService.java
@@ -3,16 +3,23 @@ package ru.forumcalendar.forumcalendar.service.base;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import ru.forumcalendar.forumcalendar.domain.Contact;
+import ru.forumcalendar.forumcalendar.domain.ContactType;
 import ru.forumcalendar.forumcalendar.domain.Role;
 import ru.forumcalendar.forumcalendar.domain.User;
 import ru.forumcalendar.forumcalendar.exception.EntityNotFoundException;
+import ru.forumcalendar.forumcalendar.model.form.ContactForm;
 import ru.forumcalendar.forumcalendar.model.form.UserForm;
+import ru.forumcalendar.forumcalendar.repository.ContactRepository;
+import ru.forumcalendar.forumcalendar.repository.ContactTypeRepository;
 import ru.forumcalendar.forumcalendar.repository.RoleRepository;
 import ru.forumcalendar.forumcalendar.repository.UserRepository;
 import ru.forumcalendar.forumcalendar.service.UploadsService;
 import ru.forumcalendar.forumcalendar.service.UserService;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 @Service
@@ -20,6 +27,8 @@ public class BaseUserService implements UserService {
 
     private final UserRepository userRepository;
     private final RoleRepository roleRepository;
+    private final ContactRepository contactRepository;
+    private final ContactTypeRepository contactTypeRepository;
 
     private final UploadsService uploadsService;
 
@@ -27,11 +36,14 @@ public class BaseUserService implements UserService {
     public BaseUserService(
             UserRepository userRepository,
             RoleRepository roleRepository,
-            UploadsService uploadsService
+            ContactRepository contactRepository,
+            ContactTypeRepository contactTypeRepository, UploadsService uploadsService
     ) {
 
         this.userRepository = userRepository;
         this.roleRepository = roleRepository;
+        this.contactRepository = contactRepository;
+        this.contactTypeRepository = contactTypeRepository;
         this.uploadsService = uploadsService;
     }
 
@@ -96,19 +108,39 @@ public class BaseUserService implements UserService {
 
         User user = getCurrentUser();
 
-        String photo = uploadsService.upload(userForm.getPhoto(), getCurrentId())
-                .map((f) -> {
-                    if (!user.getPhoto().equals(f.getName()))
-                        uploadsService.delete(user.getPhoto());
+        if (!userForm.getPhoto().isEmpty()) {
+            String photo = uploadsService.upload(userForm.getPhoto(), getCurrentId())
+                    .map((f) -> {
+                        if (!user.getPhoto().equals(f.getName()))
+                            uploadsService.delete(user.getPhoto());
 
-                    return f.getName();
-                })
-                .orElse(user.getPhoto());
+                        return f.getName();
+                    })
+                    .orElse(user.getPhoto());
+            user.setPhoto(photo);
+        }
 
         user.setFirstName(userForm.getFirstName());
         user.setLastName(userForm.getLastName());
 
-        user.setPhoto(photo);
+        if (userForm.getContactForms() != null) {
+            List<Contact> contacts = new ArrayList<>(userForm.getContactForms().size());
+
+            for (ContactForm cf : userForm.getContactForms()) {
+                Contact contact = contactRepository.findById(cf.getId()).orElse(new Contact());
+                contact.setUser(user);
+                contact.setLink(cf.getLink());
+
+                ContactType contactType = contactTypeRepository.findById(cf.getContactTypeId())
+                        .orElseThrow(() -> new EntityNotFoundException("Contact type with id " + cf.getContactTypeId() + " not found"));
+
+                contact.setContactType(contactType);
+
+                contacts.add(contact);
+            }
+
+            contactRepository.saveAll(contacts);
+        }
 
         return userRepository.save(user);
     }
